+++ utilities/packages/typescript/react-libs/src/util/jobWorkflowDefinitions.ts
@@ -68,7 +68,7 @@ export const REGISTER_MODEL_RUN_WORKFLOW_DEFINITION: WorkflowDefinition = {
 
 export const UPDATE_MODEL_RUN_WORKFLOW_DEFINITION: WorkflowDefinition = {
   name: "Model Run Update",
-  description: "Tracks the update of a model run",
+  description: "Tracks the update of a model run.",
   steps: [
     // Lodge and update step
     {
